[TECH](RC) refacto2: moved date rules to service + tests
